workflow(test): cleanup
