+++ .mocharc.js
@@ -1,5 +1,5 @@
 module.exports = {
-  require: "ts-node/register/files",
+  require: "ts-node/register/transpile-only",
   timeout: 20000,
   parallel: true,
 };
