+++ deploy/ConstantProductPoolFactory.ts
@@ -2,7 +2,7 @@ import { DeployFunction } from "hardhat-deploy/types";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
 const deployFunction: DeployFunction = async function ({ deployments, getNamedAccounts, ethers }: HardhatRuntimeEnvironment) {
-  console.log("Running ConstantProductPoolFactory deploy script");
+  // console.log("Running ConstantProductPoolFactory deploy script");
   const { deploy } = deployments;
 
   const { deployer } = await getNamedAccounts();
@@ -16,11 +16,11 @@ const deployFunction: DeployFunction = async function ({ deployments, getNamedAc
   });
 
   if (!(await masterDeployer.whitelistedFactories(address))) {
-    console.log("Add ConstantProductPoolFactory to MasterDeployer whitelist");
+    // console.log("Add ConstantProductPoolFactory to MasterDeployer whitelist");
     await (await masterDeployer.addToWhitelist(address)).wait();
   }
 
-  console.log("ConstantProductPoolFactory deployed at ", address);
+  // console.log("ConstantProductPoolFactory deployed at ", address);
 };
 
 export default deployFunction;
