+++ deploy/HybridPoolFactory.ts
@@ -1,12 +1,8 @@
 import { DeployFunction } from "hardhat-deploy/types";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
-const deployFunction: DeployFunction = async function ({
-  deployments,
-  getNamedAccounts,
-  ethers,
-}: HardhatRuntimeEnvironment) {
-  console.log("Running HybridPoolFactory deploy script");
+const deployFunction: DeployFunction = async function ({ deployments, getNamedAccounts, ethers }: HardhatRuntimeEnvironment) {
+  // console.log("Running HybridPoolFactory deploy script");
   const { deploy } = deployments;
   const { deployer } = await getNamedAccounts();
   const masterDeployer = await ethers.getContract("MasterDeployer");
@@ -16,11 +12,11 @@ const deployFunction: DeployFunction = async function ({
     args: [masterDeployer.address],
   });
   if (!(await masterDeployer.whitelistedFactories(address))) {
-    console.log("Add HybridPoolFactory to MasterDeployer whitelist");
+    // console.log("Add HybridPoolFactory to MasterDeployer whitelist");
     await (await masterDeployer.addToWhitelist(address)).wait();
   }
 
-  console.log("HybridPoolFactory deployed at ", address);
+  // console.log("HybridPoolFactory deployed at ", address);
 };
 
 export default deployFunction;
