+++ deploy/IndexPoolFactory.ts
@@ -1,12 +1,8 @@
 import { DeployFunction } from "hardhat-deploy/types";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
-const deployFunction: DeployFunction = async function ({
-  deployments,
-  getNamedAccounts,
-  ethers,
-}: HardhatRuntimeEnvironment) {
-  console.log("Running IndexPoolFactory deploy script");
+const deployFunction: DeployFunction = async function ({ deployments, getNamedAccounts, ethers }: HardhatRuntimeEnvironment) {
+  // console.log("Running IndexPoolFactory deploy script");
   const { deploy } = deployments;
 
   const { deployer } = await getNamedAccounts();
@@ -19,11 +15,11 @@ const deployFunction: DeployFunction = async function ({
   });
 
   if (!(await masterDeployer.whitelistedFactories(address))) {
-    console.log("Add IndexPoolFactory to MasterDeployer whitelist");
+    // console.log("Add IndexPoolFactory to MasterDeployer whitelist");
     await (await masterDeployer.addToWhitelist(address)).wait();
   }
 
-  console.log("IndexPoolFactory deployed at ", address);
+  // console.log("IndexPoolFactory deployed at ", address);
 };
 
 export default deployFunction;
