+++ package.json
@@ -25,7 +25,7 @@
     "registry": "https://npm.pkg.github.com"
   },
   "scripts": {
-    "build": "TS_NODE_TRANSPILE_ONLY=1 npx hardhat compile",
+    "build": "TS_NODE_TRANSPILE_ONLY=1 hardhat compile",
     "console": "hardhat console",
     "mainnet:deploy": "hardhat --network mainnet deploy",
     "mainnet:verify": "hardhat --network mainnet etherscan-verify --license GPL-3.0 --force-license",
@@ -50,11 +50,10 @@
     "moonbase:export": "hardhat --network moonbase export --export exports/moonbase.json",
     "arbitrum:deploy": "hardhat --network arbitrum deploy",
     "arbitrum:export": "hardhat --network arbitrum export --export exports/arbitrum.json",
-    "test": "cross-env TS_NODE_TRANSPILE_ONLY=1 hardhat test test/constant-product/_ConstantProduct.test.ts test/constant-product/*.test.ts test/deployer/*.test.ts test/library/*.test.ts test/Router.test.ts",
-    "test:ci": "cross-env REPORT_GAS=true CI=true yarn test:coverage ",
+    "test": "hardhat test test/constant-product/ConstantProduct.test.ts test/constant-product/*.test.ts test/deployer/*.test.ts test/library/*.test.ts test/Router.test.ts",
     "test:debug": "mocha --inspect-brk",
     "test:trace": "yarn test --logs",
-    "test:coverage": "cross-env NODE_OPTIONS=\"--max-old-space-size=8192\" CODE_COVERAGE=true npx hardhat coverage",
+    "test:coverage": "hardhat coverage --testfiles \"{test/constant-product/ConstantProduct.test.ts,test/constant-product/*.test.ts,test/abstract/*.test.ts,test/deployer/*.test.ts,test/library/*.test.ts,test/router/*.test.ts}\"",
     "test:gas": "cross-env REPORT_GAS=true yarn test",
     "prettier": "prettier --write 'test/**/*.{js,ts}' && prettier --write contracts/**/*.sol",
     "lint": "yarn prettier && solhint -c .solhint.json contracts/**/*.sol",
